a few warning cleanups, added JanusFury's support for saving PNG images with accurate pixel formats (including palettes and such)
git-svn-id: svn://svn.code.sf.net/p/corona/code/trunk/corona@156 5ef2176e-1c23-0410-8492-f621cc4e4028

+++ test/FileTests.cpp
@@ -62,7 +62,7 @@ FileTests::testMemoryLoads() {
     int file_size = ftell(file);
     byte* buffer = new byte[file_size];
     fseek(file, 0, SEEK_SET);
-    CPPUNIT_ASSERT(fread(buffer, 1, file_size, file) == file_size);
+    CPPUNIT_ASSERT((int)fread(buffer, 1, file_size, file) == file_size);
     fclose(file);
 
     // then create a memory file from it
